Mosaic now compiles and runs
 * Fixes for most compile-time warnings/errors
 * Make the png stuff work with modern libpng -- even if it crashes on some PNGs.
 * Most sites may not display just yet.

+++ libwww2/HTFile.c
@@ -1228,7 +1228,7 @@ forget_multi:
                     else
                         {
                         sprintf(buffer,"%s (%d bytes)", 
-                            dataptr, statbuf.st_size);
+                            dataptr, (int)statbuf.st_size);
               
                         format = HTFileFormat(dataptr, &pencoding, 
                                      WWW_SOURCE, &cmpr);
